% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl_presentation_html.R
\name{ncl_presentation}
\alias{ncl_presentation}
\title{Convert to an NCL xaringan presentation}
\usage{
ncl_presentation(...)
}
\arguments{
\item{...}{additional arguments to pass to [xaringan::moon_reader()]}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to an NCL branded xaringan
presentation.
}
\examples{
\dontrun{
library(rmarkdown)

# simple invocation
render("input.Rmd", ncl_presentation())
}
}
