% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl_theme_cols.R
\name{ncl_theme_cols}
\alias{ncl_theme_cols}
\title{NCL Theme Colours}
\usage{
ncl_theme_cols(
  ...,
  palette = c(NA, "ics", "icb", "icb_darkblue", "icb_brightblue", "icb_aquablue",
    "icb_lightblue", "icb_nhsblue", "icb_darkgrey", "icb_midgrey", "ics_purple",
    "ics_pink", "ics_limegreen", "ics_brightgreen", "ics_teal")
)
}
\arguments{
\item{...}{individual colours that you wish to get the values of}

\item{palette}{a name of a palette to select the colours of}
}
\value{
a named vector of RGB colours in hexadecimal form
}
\description{
This function returns the colours that are used in the NCL theme as
a named vector of RGB values in hexadecimal form.
}
\details{
If no arguments are passed to the function, then all of the colours are
returned. If only specific colours are required, then the names of the
colours that are required can be passed as strings to the function, and only
those colours will be returned.

If a "palette" is selected then no arguments should be passed to ...

Only one palette can be selected at a time.
}
\examples{
# show all of the colours
ncl_theme_cols()
# or, just show some colours
ncl_theme_cols("ics", "icb", "nhsblue")
# or, select a single palette
ncl_theme_cols(palette = "icb_nhsblue")
}
