% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl_theme_gradient_colour_scales.R
\name{ncl_theme_gradient_colour_scales}
\alias{ncl_theme_gradient_colour_scales}
\alias{scale_colour_ncl}
\alias{scale_color_ncl}
\alias{scale_fill_ncl}
\title{NCL Theme Gradient colour Scale}
\usage{
scale_colour_ncl(
  palette = c("ics", "icb", "icb_darkblue", "icb_brightblue", "icb_aquablue",
    "icb_lightblue", "icb_nhsblue", "icb_darkgrey", "icb_midgrey", "ics_purple",
    "ics_pink", "ics_limegreen", "ics_brightgreen", "ics_teal"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_color_ncl(
  palette = c("ics", "icb", "icb_darkblue", "icb_brightblue", "icb_aquablue",
    "icb_lightblue", "icb_nhsblue", "icb_darkgrey", "icb_midgrey", "ics_purple",
    "ics_pink", "ics_limegreen", "ics_brightgreen", "ics_teal"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_fill_ncl(
  palette = c("ics", "icb", "icb_darkblue", "icb_brightblue", "icb_aquablue",
    "icb_lightblue", "icb_nhsblue", "icb_darkgrey", "icb_midgrey", "ics_purple",
    "ics_pink", "ics_limegreen", "ics_brightgreen", "ics_teal"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The name of palette to use}

\item{discrete}{Optional: boolean to indicate that this scale is a discrete
scale. Defaults to TRUE}

\item{reverse}{Optional: boolean to reverse the direction of the scale.
Defaults to FALSE}

\item{...}{additional arguments passed to the ggplot functions}
}
\value{
a scale object to be used with a ggplot object
}
\description{
Generates a two colour gradient (low-high)
}
