% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl_theme_pal.R
\name{ncl_theme_pal}
\alias{ncl_theme_pal}
\title{Strategy Unit Theme Palettes}
\usage{
ncl_theme_pal(
  palette = c("ics", "icb", "icb_darkblue", "icb_brightblue", "icb_aquablue",
    "icb_lightblue", "icb_nhsblue", "icb_darkgrey", "icb_midgrey", "ics_purple",
    "ics_pink", "ics_limegreen", "ics_brightgreen", "ics_teal"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The name of palette to use}

\item{reverse}{Optional: boolean value to reverse the palette}

\item{...}{arguments to pass to colorRamp}
}
\value{
A colorRamp function
}
\description{
This function returns a palette generated by colorRampPalette from one of the
Strategy Unit palettes
}
\examples{
ncl_theme_pal()
# or, specify a paltte
ncl_theme_pal("ics")
# a reverse palette
ncl_theme_pal("icb", TRUE)
}
