% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl_theme.R
\name{theme_nclicb}
\alias{theme_nclicb}
\title{Creates the NCL ICB Theme for ggplot}
\usage{
theme_nclicb(base_family = "opensans")
}
\arguments{
\item{base_family}{The font to use in this theme}
}
\description{
Creates a base theme for ggplot objects. By default this uses a font that
requires the \link[extrafont]{extrafont} package to be installed.
}
\examples{
\dontrun{
library(ggplot2)
library(NCLtemplates)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  theme_nclicb()
}
}
