% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl_theme.R
\name{theme_nclics}
\alias{theme_nclics}
\title{Creates the NCL ICS Theme for ggplot}
\usage{
theme_nclics(base_family = "Mulish")
}
\arguments{
\item{base_family}{The font to use in this theme. This defaults to 'Mulish', but you can change it to any font you have installed.}
}
\description{
Creates a base theme for ggplot objects. By default this uses a font that
 requires installing, and this is handled in the package load and `ncl_setup` function package to be installed.
}
\examples{
\dontrun{
library(ggplot2)
library(NCLRtemplates)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  theme_nclics()
}
}
\seealso{
[ncl_setup()], [show_text_auto()], [font_font_add_google()], [theme_nclicb()], [theme()], [theme_classic()]
}
