% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl_theme.R
\name{theme_nclicb}
\alias{theme_nclicb}
\title{Creates the NCL ICB Theme for ggplot}
\usage{
theme_nclicb(base_family = "Mulish")
}
\arguments{
\item{base_family}{The font to use in this theme}
}
\description{
Creates a base theme for ggplot objects. By default this uses a font that
requires installing, and this is handled in the package load and `ncl_setup` function package to be installed.
}
\examples{
\dontrun{
library(ggplot2)
library(NCLtemplates)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  theme_nclicb()
}
}
\seealso{
[ncl_setup()], [show_text_auto()], [font_font_add_google()], [theme_nclics()], [theme()], [theme_classic()]
}
